added read timeout on okhttpclient, created application nnavigation file and applied some improvments in country info screen and counntry info list files
